ada: Avoid failing reanalysis of aggregate bounds
An expression like Some_Package."+" (Arg1, Arg2) is sometimes transformed
during expansion into an unqualified call of the form "+" (Arg1, Arg2).
This is normally ok, but it means that reanalysis of the expression is likely
to fail and must therefore be avoided. Remove code that was forcing such
reanalysis (the original reason for the now-removed code is unknown).

gcc/ada/ChangeLog:

	* exp_aggr.adb (Check_Bounds): Remove calls that were setting the
	Analyzed flag to False for two aggregate bound expressions.