replace atoi with strtoul in varasm.cc (decode_reg_name_and_count) [PR114540]
The function uses atoi, which can silently return valid numbers even for
some too large numbers in the string.

Furthermore, the verification that all the characters in asmspec are
decimal digits can be simplified when using strotoul, we can check just
the first digit and whether the end pointer points to '\0'.

2024-12-03  Heiko Eißfeldt  <heiko@hexco.de>

	PR middle-end/114540
	* varasm.cc (decode_reg_name_and_count): Use strtoul instead of atoi
	and simplify verification that the whole asmspec contains just decimal
	digits.

	* gcc.dg/pr114540.c: New test.

Signed-off-by: Heiko Eißfeldt <heiko@hexco.de>
Co-authored-by: Jakub Jelinek <jakub@redhat.com>