aarch64: Fix fp8 cpuinfo feature names
The previous version of the patch was based on the mistaken assumption that
features in /proc/cpuinfo had matching names to the feature names that gcc and
gas accept.
This patch enables the fp8 feature when the f8cvt feature is enabled, under the
assumption that fpmr is always enabled when f8cvt is.

Changelog:

gcc/
	* config/aarch64/aarch64-option-extensions.def: (fp8): fix FEATURE_STRING.
	(fp8fma, ssve-fp8fma): Likewise.
	(fp8dot4, ssve-fp8dot4, fp8dot2, ssve-fp8dot2): Likewise.