tree-optimization/116083 - SLP discovery slowness
One large constant factor of SLP discovery is figuring the vector
type for each individual lane of each node.  That should be redundant
since the structual comparison of stmts should ensure they end up
the same so the following computes them only once per node rather
than for each lane.

This cuts the compile-time of the testcase in half.

	PR tree-optimization/116083
	* tree-vect-slp.cc (vect_build_slp_tree_1): Compute vector
	type and max_nunits only once.  Remove check for matching
	vector type of each lane and replace it with matching check
	for LHS type.