cfgexpand: Reverse the order of going through the update_cache_list queue.
This is a small optimization, the reversed order of the walk of update_cache_list queue.
The queue is pushed in Pre-order/NLR, reversing the order will reduce how many times we
need to go through the loop as we update the nodes which might have a link back to another
one first.

Bootstrapped and tested on x86_64-linux-gnu.

gcc/ChangeLog:

	* cfgexpand.cc (vars_ssa_cache::operator()): Reverse the order of the going
	through the update list.

Signed-off-by: Andrew Pinski <quic_apinski@quicinc.com>