c++: Fix up erroneous template error recovery ICE [PR117826]
The testcase in the PR (which can't be easily reduced and is
way too large and has way too many errors) results in an ICE,
because the erroneous_templates hash_map holds trees of erroneous
templates across ggc_collect and some of the templates in there
could be removed, so the later lookup can crash on comparison of
already freed and reused trees.

The following patch makes the hash_map GTY((cache)) marked.
The cp-tree.h changes before the erroneous_template declaration
are needed to make gengtype happy, it didn't like using
directive nor using a template-id as a template parameter.

It is marked cache because if a decl would be solely referenced from
the erroneous_templates hash_map, then nothing would look it up.

2024-12-04  Jakub Jelinek  <jakub@redhat.com>

	PR c++/117826
	* cp-tree.h (struct decl_location_traits): New type.
	(erroneous_templates_t): Change using into typedef.
	(erroneous_templates): Add GTY((cache)).
	* error.cc (cp_adjust_diagnostic_info): Use
	hash_map_safe_get_or_insert<true> rather than
	hash_map_safe_get_or_insert<false> for erroneous_templates.